bi_agent/requirements.txt
anthropic>=0.49.0
httpx>=0.27.0
slack-bolt>=1.21.0
slack-sdk>=3.33.0
duckdb>=1.2.0
pyyaml>=6.0.2
matplotlib>=3.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
